♻ Make analyze task privy to our tools
