+++ platform/reworkd_platform/tests/agent/agent/tools_test.py
@@ -1,4 +1,4 @@
-from reworkd_platform.web.api.agent.tools.think import Think
+from reworkd_platform.web.api.agent.tools.reason import Reason
 from reworkd_platform.web.api.agent.tools.tools import (get_tools_overview,
                                                         get_tool_from_name,
                                                         get_default_tool, get_tool_name)
@@ -7,7 +7,7 @@ from reworkd_platform.web.api.agent.tools.wikipedia_search import Wikipedia
 
 def test_get_tool_name() -> None:
     assert get_tool_name(Wikipedia) == "wikipedia"
-    assert get_tool_name(Think) == "think"
+    assert get_tool_name(Reason) == "think"
 
 
 def test_get_tools_overview() -> None:
@@ -20,9 +20,9 @@ def test_get_tools_overview() -> None:
 def test_get_tool_from_name() -> None:
     assert get_tool_from_name("Wikipedia") == Wikipedia
     assert get_tool_from_name("WiKiPeDia") == Wikipedia
-    assert get_tool_from_name("Think") == Think
-    assert get_tool_from_name("NonExistingTool") == Think
+    assert get_tool_from_name("Think") == Reason
+    assert get_tool_from_name("NonExistingTool") == Reason
 
 
 def test_get_default_tool() -> None:
-    assert get_default_tool() == Think
+    assert get_default_tool() == Reason
