+++ platform/reworkd_platform/web/api/agent/agent_service/open_ai_agent_service.py
@@ -1,52 +1,19 @@
 from os import environ
-from random import randint
 from typing import List, Optional
 
 from langchain.chains import LLMChain
-from langchain.chat_models import ChatOpenAI
 
-from reworkd_platform.settings import settings
 from reworkd_platform.web.api.agent.agent_service.agent_service import AgentService
 from reworkd_platform.web.api.agent.analysis import Analysis, get_default_analysis
 from reworkd_platform.web.api.agent.helpers import extract_tasks
-from reworkd_platform.web.api.agent.model_settings import ModelSettings
+from reworkd_platform.web.api.agent.model_settings import ModelSettings, create_model
 from reworkd_platform.web.api.agent.prompts import (
     start_goal_prompt,
     analyze_task_prompt,
     execute_task_prompt,
     create_tasks_prompt,
 )
-
-GPT_35_TURBO = "gpt-3.5-turbo"
-
-
-def get_server_side_key() -> str:
-    keys = [
-        key.strip() for key in (settings.openai_api_key or "").split(",") if key.strip()
-    ]
-    return keys[randint(0, len(keys) - 1)] if keys else ""
-
-
-def create_model(model_settings: Optional[ModelSettings]) -> ChatOpenAI:
-    _model_settings = model_settings
-
-    if not model_settings or not model_settings.customApiKey:
-        _model_settings = None
-
-    return ChatOpenAI(
-        openai_api_key=_model_settings.customApiKey
-        if _model_settings
-        else get_server_side_key(),
-        temperature=_model_settings.customTemperature
-        if _model_settings and _model_settings.customTemperature is not None
-        else 0.9,
-        model_name=_model_settings.customModelName
-        if _model_settings and _model_settings.customModelName is not None
-        else GPT_35_TURBO,
-        max_tokens=_model_settings.maxTokens
-        if _model_settings and _model_settings.maxTokens is not None
-        else 400,
-    )
+from reworkd_platform.web.api.agent.tools.tools import get_tools_overview
 
 
 class OpenAIAgentService(AgentService):
@@ -65,9 +32,11 @@ class OpenAIAgentService(AgentService):
     ) -> Analysis:
         llm = create_model(model_settings)
         chain = LLMChain(llm=llm, prompt=analyze_task_prompt)
-        actions = ["reason", "search"]
 
-        completion = chain.run({"goal": goal, "task": task, "actions": actions})
+        completion = chain.run(
+            {"goal": goal, "task": task, "tools_overview": get_tools_overview()}
+        )
+
         print("Analysis completion:\n", completion)
         try:
             return Analysis.parse_raw(completion)
