+++ platform/reworkd_platform/web/api/agent/prompts.py
@@ -14,12 +14,12 @@ start_goal_prompt = PromptTemplate(
 analyze_task_prompt = PromptTemplate(
     template="""You have the following higher level objective "{goal}". You currently
     are focusing on the following task: "{task}". Based on this information, evaluate
-    what the best action to take is strictly from the list of actions: {actions}. You
-    should use 'search' only for research about current events where "arg" is a
-    simple clear search query based on the task only. Use "reason" for all other
-    actions. Return the response as an object of the form {{ "action": "string",
-    "arg": "string" }} that can be used in JSON.parse() and NOTHING ELSE.""",
-    input_variables=["goal", "actions", "task"],
+    what the best action to take is strictly from the list of actions below:\n\n
+    {tools_overview}\n\n
+    Return your response in an object of the form\n\n
+    {{ "action": "string","arg": "string" }}\n\n
+    that can be used in JSON.parse() and NOTHING ELSE.""",
+    input_variables=["goal", "task", "tools_overview"],
 )
 
 execute_task_prompt = PromptTemplate(
