+++ platform/reworkd_platform/web/api/agent/tools/utils.py
@@ -2,10 +2,7 @@ from typing import List
 
 from langchain import LLMChain
 
-from reworkd_platform.web.api.agent.agent_service.open_ai_agent_service import (
-    create_model,
-)
-from reworkd_platform.web.api.agent.model_settings import ModelSettings
+from reworkd_platform.web.api.agent.model_settings import ModelSettings, create_model
 from reworkd_platform.web.api.agent.prompts import summarize_prompt
